fix(types): move resolveBrowserLocale out of layout into src/lib/locale
Next.js typed routes forbid extra exports from layout files, so
exporting resolveBrowserLocale from the (frontend) layout made
`tsc --noEmit` fail on the generated .next/types validator. Move it
to a dedicated server-side module (same pattern as src/lib/domain.ts)
and update all importers.
