+++ src/app/(frontend)/layout.tsx
@@ -4,14 +4,10 @@ import theme from "@/theme/theme";
 import { ThemeProvider } from "@mui/material";
 import { Inter } from "next/font/google";
 import type { Metadata } from "next";
-import { headers } from "next/headers";
 import { getDomain } from "@/lib/domain";
+import { resolveBrowserLocale } from "@/lib/locale";
 import { getTenantBySubDomain } from "@/lib/data/tenants";
-import {
-  PAYLOAD_SUPPORTED_LOCALES,
-  type PayloadLocale,
-  resolveTenantLocale,
-} from "@/utils/locales";
+import { resolveTenantLocale } from "@/utils/locales";
 
 const inter = Inter({
   subsets: ["latin"],
@@ -27,34 +23,6 @@ export const metadata: Metadata = {
     icon: "/favicon.ico",
     shortcut: "/favicon.ico",
   },
-};
-
-export const resolveBrowserLocale = async (): Promise<PayloadLocale> => {
-  const headerList = await headers();
-  const acceptLanguage = headerList.get("accept-language");
-  if (!acceptLanguage) {
-    return "en";
-  }
-
-  const supported = new Set<PayloadLocale>(PAYLOAD_SUPPORTED_LOCALES);
-  const candidates = acceptLanguage
-    .split(",")
-    .map((entry) => entry.trim().split(";")[0]?.toLowerCase())
-    .filter(Boolean);
-
-  for (const candidate of candidates) {
-    if (!candidate) continue;
-    if (supported.has(candidate as PayloadLocale)) {
-      return candidate as PayloadLocale;
-    }
-
-    const [base] = candidate.split("-");
-    if (base && supported.has(base as PayloadLocale)) {
-      return base as PayloadLocale;
-    }
-  }
-
-  return "en";
 };
 
 export default async function RootLayout(props: { children: React.ReactNode }) {
